Save the npz dtype as bytes
Currently, the npz reader expects the dtype to be in bytes
([here](https://github.com/joelvbernier/hexrd3/blob/2f841ffb7780a74af4f798c08aadd4a9160e5cbd/hexrd/imageseries/load/framecache.py#L68)).
The npz writer was not writing in bytes, which resulted in an
incompatibility between the two.

This commit modifies the writer to write out the npz dtype
in bytes so that the reader is able to read it correctly. It
fixes the incompatibility between the two.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>