Add missing del_eta to make_powder_rings
Most of the time, when a user would call `make_powder_rings()`,
they probably passed a PlaneData object. However, the function is
also designed to accept a list of two theta values. It appears,
though, that if you pass it a list of two theta values, it misses
the creation of del_eta that is normally done if a PlaneData object
is passed in.

This adds in the missing variable. The function seems to work properly
if this variable is included.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>