Remove unused variable
In PlanarDetector.make_powder_rings(), if the material's plane data
has a tThWidth of None, and no delta_tth is passed to the function,
an exception occurs when `np.degrees(pd.tThWidth)` is attempted.

Upon looking further, though, it appears that the `delta_tth`
variable is not used beyond that point, so it seems to be not
necessary to set it.

Remove the variable so that the exception goes away.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>