Move code to create clustering params into method
This moves the majority of the code used to create the min samples
and the mean reflections per grain, for the clustering, into
a separate function. This is done so that we can call this function
from the GUI.

I did a CLI `hexrd find-orientations ...` before and after test,
to make sure the output didn't change, and a diff shows that the
output appears to be the same.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>