Make the compliance tensor a computed property
Rather than storing both the stiffness and compliance tensors, which
requires storing two tensors, and can run the risk of one being out of sync
with the other, make the compliance tensor a computed property.

This way, the only matrix that is stored is the stiffness tensor. If one
tries to get the compliance via `unitcell.compliance`, it will be computed
and returned. If one sets the compliance tensor via `unitcell.compliance = t`,
it will be converted to stiffness and stored as the stiffness tensor.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>