Use 'cache' and 'nogil' for extract_ijv numba
Caching will cause the compiled functions to be cached in the
filesystem, and speed up future uses of it.

'nogil' gets numba to release the GIL.

Both of these seem to provide a nice speed-up for multi-threading.
The numba function is still a rate-limiting step, however, because
the numba compiler has a lock and only one thread can compile
at a time.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>