Check if file is writable with 'r+' instead of 'w'
On Windows, checking if a file is writable via
`os.access(__file__, os.W_OK)` unfortunately does not work. We have
had to resort to a more brutish method of trying to open the file, like
this:
```python
def is_writable_file(path):
    try:
        with open(path, 'w'):
            return True
    except Exception:
        return False
```

Unfortunately, just the act of opening a file with 'w' erases its
contents. We need to open with 'r+' instead to avoid this.

What the previous opening with 'w' resulted in:

1. Hexrd on Windows installed via a package would work correctly,
   because the user running "hexrd" doesn't have permissions to write
   to the file (since the package is installed with admin privileges).
2. Hexrd on Windows installed via conda/pip would cause a bug to occur,
   because the user running "hexrd" does have permissions to write to
   the file, and opening with "w" would erase its contents.

Opening with 'r+' also raises an exception if the file is not
writable, so this change fixes the bug and still checks if a file is
writable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>