Copy old PlaneData attributes to new PlaneData
When we have to create a new PlaneData object because the HKLs have
been modified, copy over the old PlaneData attributes to the new one.

This also computes a mapping between the old HKLs and the new ones,
and it sets the exclusions of the new PlaneData object from the old
one via the mapping. If all of the HKLs are excluded, it just uses
the default ones.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>