Use numba, if possible, in hexrd.symmetry.isnew
Profiling hexrdgui indicated that the most time consuming function
on hexrdgui startup is hexrd.symmetry.isnew. Adding a numba
decorator provides a 20-40% speed-up for starting hexrdgui when
using numba.

This commit also adds and uses a decorator in hexrd.utils.decorators
called "numba_njit_if_available". Basically, if we are using numba,
it forwards the decorator arguments to numba.njit. And if we are not
using numba, it does nothing.

If we want to improve hexrdgui start-up time in the future, profiling
indicates that these are the next most time consuming functions (in
order):

```python
hexrd.unitcell.unitcell.CalcLength
hexrd.matrixutil.findDuplicateVectors
hexrd.unitcell.unitcell.CalcStar
```

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>