Perform multiprocessing on pixel_solid_angles
Threading was attempted, but it appeared that all the threads would
either be fighting for the GIL or fighting for the compiler lock in
numba. Threading actually resulted in a large slowdown.

Multiprocessing seems to work well, since there is not a lot of
data to transmit back and forth. For 2048 rows by 2048 columns
and pixel sizes of 0.2, I was able to go from about 46 seconds
to about 6.8 seconds for running this property.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>