Use numba.njit(cache=True, nogil=True) for all
Using "njit" instead of "jit" means to use the "nopython" argument,
which can be faster. But it may not be allowed for some function
argument types.

Caching will get numba to cache the functions it compiles so that it
does not have to compile them again. This can provide a big speed-up
after the functions have been called once and saved.

Note that if you edit the file that contains the function, the cache
will automatically be re-generated.

nogil tells numba to release the GIL. This is helpful in case the
function is used in any python multi-threading.

Caching is also helpful if the function is used in any python
multi-threading because numba can only compile on one thread at a time...

In any case, these changes all have the potential to be good, but we should
definitely test them all, and make sure that everything still works, gives
the same results, and takes at most the same amount of time that it did
previously.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>