Improve performance of eta omega map generation
This generates ring parameters one time in its own loop, and then
swaps the previous ordering of the loops so that looping over the
images is on the outside and looping over the rings is on the inside.

This results in each image only being loaded once, rather
than once for each ring, as it was before. Since loading the images is
time consuming, this results in a nice speedup.

For the GE example in the examples repository, my performance improved
from 17 seconds to 5 seconds, and the same result was obtained.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>