Add max_workers kwarg to HEDMInstrument.__init__()
This is used in the HEDMInstrument object and forwarded to the
PlanarDetector objects as well. In the future, we can consider
splitting it among the detectors rather than just forwarding it.

Also, now the "multiprocessing" configure option in the HEDM config
object is passed to the HEDMInstrument object when it gets created,
so they should match.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>