Ensure root log level gets set
On some systems, setting the root log level manually is not required,
but it gets set via the basicConfig().

However, on the cluster I am working with, the root log level does
not get set via basicConfig(), and I must set it manually.

Setting it manually like this seems to fix the issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>