Force upload packages to anaconda
Before this commit, if we re-run the GitHub actions workflow (which may
be done if a workflow failed for a network issue, for instance), the
workflow will fail when it tries to upload the anaconda package because
the anaconda package already exists on anaconda.

Force the upload so that this error does not happen, and so that the
packages on anaconda are generated from the latest workflow that was ran.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>