Add a maxsize argument to memoize with LRU removal
If the maxsize of the cache is about to be exceeded, the least
recently used item in the cache will be removed.

The maxsize defaults to 5. We can change it if needed.

This also adds a `cache_info()` function similar to `functool.lru_cache()`'s
`cache_info()` function, which returns hits, misses, current size and max size
of the memoized function.

Tests were also added that seem to confirm that the decorator works properly.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>