Revert #253
@saransh and I spoke about this. We agreed that the previous behavior
should be restored, so that exclusions do not get changed when
material properties get modified (such as is done sometimes with WPPF,
or when the user modified material settings in the GUI).

He will also track down and fix the specific issue he was encountering
that caused him to submit #253.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>