Convert to strings for h5py string types
In h5py >= 3.0.0, h5py no longer automatically converts to string
data types, but always leaves string/byte data as bytes. Thus, we
need to add some compatibility to automatically convert to strings
when the underlying data is a string.

This fixes two specific cases that I currently know of: #273 and
loading in a distortion string in an instrument hdf5 file.

We should add this in more places if I missed any.

Fixes: #273

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>