Use option to return spot data from pull_spots()
The previous `save_spot_list` argument was intended to be used for
this purpose, but it was unused.

This was renamed to `return_spot_list` to emphasize that the spot
list will be returned (rather than some other form of saving such
as writing to disk). And now, if `return_spot_list` is `True`, the
`patch_data` will be appended to the data that is returned.

So far, I believe this is the only data needed right now for
visualizing the spots data.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>