Return all spots data for "return_spot_list"
Previously, we were just returning the patch data. However, it
appears that the spot montages require several more pieces of data
that were not included, namely: tth_edges, eta_edges, and ome_centers.

Instead of picking and choosing the pieces of data, which may change
depending on what we want to plot in the montages, let's just return
all of the data...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>