Fix pickling for ProcessedImageSeries
From the [pickle documentation:](https://docs.python.org/3.8/library/pickle.html#what-can-be-pickled-and-unpickled)
```
Similarly, when class instances are pickled, their class’s code and data
are not pickled along with them. Only the instance data are pickled.
```

Class data does not get pickled, and the class's `__init__()` method
also does not get called when the object is unpickled. This causes
the `_opdict` class variable in `ProcessedImageSeries` to be empty
when it gets pickled and unpickled [here](https://github.com/HEXRD/hexrd/blob/408bea18f11f3ad968bbefa64518f530bf0882dd/hexrd/fitgrains.py#L83) for when
multiprocessing uses spawning rather than forking (and python 3.8
defaults to use forking only on Linux).

Thus, to make sure the `_opdict` gets pickled as well, we can make it
a member attribute rather than a class attribute.

Another alternative would be to add functions that get called any
time the `_opdict` is to be accessed that make sure it is populated
with the default opts, but I believe making `_opdict` a member
attribute is cleaner.

The issue of the `_opdict` being empty would specifically be
encountered during fit grains if the OS is not Linux, and there
are at least 2 grains (which results in multiprocessing being
used).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>