Resize instrument memoization funcs on __init__
When an HEDMInstrument is initialized, it now checks to see if the
memoization functions it uses have a large enough size for each
detector. If they do not, it resizes them.

This should hopefully allow us to utilize the memoization for
large numbers of detectors while avoiding having too big of a
cache size for a few (but very large) detectors.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>