Remove _memo_hkls dict and its memoization
This has caused many hard-to-track bugs in hexrdgui, because it only
memoizes based upon the `PlaneData` object, and does not check to see
if the exclusions or any other properties of the `PlaneData` object has
changed before returning its memoized result. This results in incorrect
hkls being returned when some of the `PlaneData` properties have been
changed.

It does not appear to make much of a difference in timings. For the
`NIST_ruby` example, I get these timings:

```
Without memo hkls:
find-orientations: 17.4 s
fit-grains: 112.5 s

With memo hkls:
find-orientations: 18.3 s
fit-grains: 110.5 s
```

So we can just remove it rather than improving the memoization.

It also does not need to return a contiguous array of floats, so we can
remove those parts.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>