Fix conda numpy version to 1.20
The build step of hexrd was using numpy 1.22, because it is the latest.
However, at runtime, numpy 1.20 would be installed because it is the
latest version that numba supports.

The mismatch in version caused an error containing the following:
```python
RuntimeError: module compiled against API version 0xf but this version of numpy is 0xe
```

Fix the versions of numpy to 1.20 so they will both match. We can bump
the numpy version when numba supports newer versions.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>