Set default dmin to None in load_materials_hdf5()
For the dmin argument, `Material.__init__()` treats `None` and
`Material.DFLT_DMIN` differently. If it is `None`, then the dmin is read
from the HDF5 file if available. However, if it is not `None`, as was
the case before, then the dmin will not be read from the HDF5 file.

Default it to `None` in `load_materials_hdf5()` so that it will be read
from the HDF5 file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>