Perform eta shift to get the correct values
It appears that Ryan's example had eta shifted by a certain amount
compared to what we use in HEXRD. And the shift for TARDIS and PXRDIP
are different.

Perform these shifts so we get the correct answer.

This also allows us to compute the pinhole distortion on multiple panels
at once, and adds a function for generating the pinhole distortion for
a whole polar field.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>