A few final fixes for pinhole distortion
We were accidentally modifying eta in place, which meant that repeated
calls would give different values. That is now fixed.

We also flipped the return of the polar tth correlation map to have the
correct sign.

We also are now ignoring the numpy warning that was coming out of the
`qq_p` computation since that is expected.

Everything now appears to be working correctly in the GUI.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>