Improve performance of Rygg pinhole method
First of all, this avoids creating the 4D arrays that were being created
before. They are avoided by looping over the columns and running them
one at a time. We get the same result, and it uses *significantly* less
memory (and time) for large angular grids.

Afterwards, it was found that for angular grids, using a thread pool and
numba provided a great speed improvement of ~6x.

However, for single columns (such as overlays), using a thread pool
and numba both appeared to slow it down.

So this adds the speed-up code for angular grids, while still performing
the serial non-numba version for overlays.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>