Break up instrument into separate modules
The instrument file was getting too big. Since we are adding new classes
to it anyways, go ahead and break it up into separate files. We are still
exposing the same classes and functions that we were exposing before (as
far as I can see, but we can add any if missing).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>