Demonstrate differences in gvec_to_xy
This commit provides a demonstration of the new differences in the
output of gvec_to_xy. It records (and prints) the max difference
in the outputs from the old xfcapi and the new xfcapi.

We had to go back and separate out the array and non-array versions
of the functions to make sure we called the correct old xfcapi version
for comparison.

We ultimately see a max difference in the outputs of the functions of
about `8e-13` (also, the shapes are sometimes slightly different as well,
where sometimes the old function returned a shape of `(1, 2)`, while the
new function would return a squeezed shape of `(2,)`. However, this did
not appear to make any difference in the output of fit-grains).

Right now, the old output is being returned, and we can see that we
get exactly the same numbers in the `grains.out` file as the old xfcapi,
which shows that `gvec_to_xy` is the source of the differences.

If the new output is returned instead, we get the new answers in the
`grains.out` file, where all numbers except completeness differ around
`1e-5` to `1e-8`.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>