Fix hexrd.material circular import
If you start a program by running `from hexrd import material`, this triggers
a circular import, because the material ends up triggering an import of the
instrument, which triggers an import of `hexrd.material` again.

Do not import the instrument in `xrdutils`. Let's keep it simple. This fixes
the circular import.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>