Use latest lxml
Fabio uses lxml (it is optional, but installing fabio installs it by
default).

On Windows, lxml 4.9.1 is causing the application to crash silently.
There are no errors - the application just crashes when fabio is imported.

We need to force it to use lxml 4.9.2 or greater to avoid this crash. We
could alternatively just stop it from being installed as well if we run
into more issues in the future.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>