Add ability to cancel fit-grains programmatically
This adds an optional argument to `fit_grains()` named
`check_if_canceled_func`. It should be a nullary function
that returns a boolean.

If provided, it will be checked every 0.25 seconds while fit-grains
is running to see if the operation should cancel.

If it returns `True`, the multiprocessing pool is terminated and
`None` is returned.

Because we rely on the multiprocessing pool to cancel, if the cancel
function is provided, we use the multiprocessing pool, even if there
is only one process to run.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>