Revive the HEXRD readthedocs documentation
A lot has changed since the documentation was last generated 6 years ago,
including sphinx and its extensions. Automatic documentation is now simpler,
so a lot of files previously present were no longer needed and are deleted in
this commit.

The process of generating the documentation is now quite simple. See the
`docs/README.md` file for more info.

This overall seems to generate the documentation correctly.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>