Fix branch cut logic for smaller eta steps
Smaller values in eta step were breaking the branch cut logic here.

The problem is that the arc length was not always less than 1e-4 when
the eta_step was small (such as 0.05). It was less than 1e-2, so still
close to zero, but not small enough.

Then on top of that, some of the ring break logic was not working.

To fix both of those issues, we loosen the logic to instead check to
see if nearly the whole two pi range is covered, and if there is a big
gap. If there is, then we assume it is a branch cut. And we locate the
break via the big gap.

This worked properly for eta step values from 0.01 to 1

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>