Install pybind11 if it is missing
We currently support 3 different build setups:

1. In a conda environment (uses pybind11 conda package in build step)
2. In a regular pip environment (uses pybind11 in build isolation as specified in pyproject.toml)
3. In a conda development environment (no build isolation)

This PR fixes number 3.

pybind11 is not a runtime dependency for the conda environment, so if a user
installs the HEXRD prerelease, it won't come with pybind11. Then, if a user
tries to make a dev environment, pybind11 won't be available.

The changes here make it so that in situation number 3, the pybind11 source
code is downloaded automatically and installed, so that the headers can
be found.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>