Fix integer overflow issue in fabio
fabio would perform some arithmetic with the frame number we provided
it to compute byte offsets for GE images. If our frame number was
np.uint32, for large datasets, this would cause an overflow error,
since uint32 would be used for the byte offset.

Instead, ensure a python integer is passed, which *cannot* overflow.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>