Invalidate sample layer past critical beta
xy points on the detector past the critical beta have very high values
and start appearing unrealistic. Just set any values to nan that are past
the critical beta.

Note that we are not invalidating map values past the critical beta for
now, because they actually appear okay when warping the polar view. This
is because they move too far and disappear (which is actually what we
want when warping to the polar view).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>