Allow Euler convention setting for calibration
Some users strongly prefer to use a different Euler convention for
calibration, such as extrinsic XYZ. It is easy to convert angles from
one Euler convention to another. However, it is difficult to convert
boundary conditions (min/max ranges for each angle) from one convention
to another.

So we add the ability to choose the Euler convention to use in calibration.
This allows the min/max ranges to be specified in lmfit using that convention.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>