Add eiger-stream-v1 imageseries format
This supports our current eiger stream data files.

It would be really nice in the future if we figured out a way to get
the data to just be regular datasets in the HDF5 file that are automatically
decompressed on loading. But we will have to do this for now...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>