Add an imageseries adapter for arbitrary functions
This adds a highly customizable imageseries adapter that simply
accepts a function and a number of frames. The function takes a
single argument (an integer for the frame index), and returns a
2D numpy array for that frame.

This will be used for streaming in data, where the data is compressed
in memory (not on disk), and the custom function simply decompresses
the frame and returns it.

We will then use this to create a sparse data writer.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>