Save eiger metadata as a yaml string
Because the numpy frame-cache format does not yet support nested
metadata structures, we cannot save nested metadata as a frame-cache
without causing errors.

We have some potential solutions to support nested metadata on the
horizon. Until that time, let's store the eiger stream metadata as
a yaml string, so that sparse arrays can actually be created from it.
The metadata is still present, it is just in the form of a yaml string.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>