Add relative constraints for whole system
The relative constraints for the whole system require the detectors to
all be translated and rotated together about the center of the system.

This includes a fairly thorough test.

It won't be that difficult to add relative constraints for groups too,
but that should be done in a separate branch.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>