Add multi-xrs support to the instrument and powder
This adds support for multiple x-ray sources to the instrument
and powder calibration. This will be the base for all multi-xrs
work going forward.

Most instruments will not be using multiple x-ray sources, so a
concept of an "active beam" is present so that software can assume
there is only one beam. Properties like `instr.beam_energy` return
the energy of the active beam.

Another method was also added to easily switch between beam sources
when needed. The `with_xray_source()` context manager can be used to
temporarily switch the active beam to another beam, if some calculations
that assume only one beam need to be performed with a different active
beam. This results in significantly cleaner code than if we had to
make all hexrd source code multi-beam aware.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>