Add several speedups for polar view generation
Perhaps the most important speedup added here is the ability to generate
and cache a mapping of the xy coordinates on each detector to the
angular pixels in the polar view. This is extremely valuable when you
must process many images that have identical instrument configurations,
since you can generate this mapping and re-use it on all subsequent
polar view generations.

This also implements several other speedups to polar view generation
as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>