FrameCacheImageSeries: remove non-pickleable attributes from state
The multiprocessing library uses pickle to serialize objects between processes. This
was creating an issue for FrameCacheImageSeries since it contained a non-pickable attribute.
We now skip this from the state. This should not affect the functionality of the framecache since
it was used to make sure that only one thread was populating the cache at a time. The same is true
now. In every process only one thread at a time can populate the cache