Only pass one rMat_s to angularPixelSize
We were previously passing an array of matrices, and only the first matrix
was being used in the old transforms code. In the new transforms code, an
error is being raised because it expects only a single matrix (and not an array
of matrices). We should explicitly only pass this first matrix to
`angularPixelSize()`.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>