Don't modify "vary" on params with an expression
The expression indicates that the parameter is computed, and thus should
not be allowed to vary.

This prevents an issue where, if the Y value of 'IMAGE-PLATE-4' for
TARDIS was allowed to vary using the old parameter flags, and TARDIS
engineering constraints are used such that the Y value of 'IMAGE-PLATE-4'
is computed, that we won't accidentally set that parameter to vary.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>