Disable engineering constraints in test
Due to a bug where we were accidentally setting IP4-Y to vary, even
though it is supposed to be computed, we were actually not applying
engineering constraints in this test.

Since we weren't actually applying engineering constraints in the
testing before, just disable them for now. We can look into adding
them later.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>