Mask out invalid pixels in the polar view
This makes it so that any invalid pixels (as defined by the detector's
panel buffer) do not have an effect on the resulting polar image and
lineout. Any polar view pixels that were interpolated with an invalid
pixel will be nan.

The intended usage is as follows:

```python
img = pv.warp_image(image_dict, pad_with_nans=True,
                    do_interpolation=True)
cake = np.mean(img, axis=0)
```

The `img` that is returned from the warp image is a masked array,
with all invalid pixels being masked out. If `np.mean(img, axis=0)` is
applied to this, it ignores all masked out pixels when computing the
mean along the axis (which is the behavior we want). This is preferred
over doing something like `np.sum(img.filled(0), axis=0)`, since columns
with invalid pixels will be reduced in value.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>