Return numpy dtype from eiger-stream-v1
All of the other imageseries adapters return a numpy dtype. Here,
we were just returning a string representing a dtype. However, we
should return a numpy dtype to be consistent with all of the other
imageseries adapters as well.

This also fixes issues in other places in the code where it is
assumed that a numpy dtype is returned from the imageseries.

Fixes: hexrd/hexrdgui##1780

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>