Refactor HEDM calibration
This refactors the HEDM calibration to work in our newer calibration
workflow, which includes support for the more sophisticated lmfit
fitting tools, including support for boundary constraints and
relative constraints between detectors, or subpanels within a larger
detector.

With this change, the old calibration methods are completely unused
and were thus removed.

A test was also added that verifies the HEDM calibration is running
successfully.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>